Type-casted argument to sprintf so that it works properly

+++ mpimatmult1.c
@@ -77,9 +77,9 @@ MPI_Bcast(matrix2, mat2size, MPI_INT, source, MPI_COMM_WORLD);
 
 if( rank == source )
 {
-sprintf(outstring,"Matrix 1 memory usage per process: %f bytes \n",mat1size*sizeof(int)/size);
+sprintf(outstring,"Matrix 1 memory usage per process: %f bytes \n",(float) mat1size*sizeof(int)/size);
 debugprintf(outstring);
-sprintf(outstring,"Matrix 2 memory usage per process: %f bytes \n",mat2size*sizeof(int));
+sprintf(outstring,"Matrix 2 memory usage per process: %f bytes \n",(float) mat2size*sizeof(int));
 debugprintf(outstring);
 }
